Support common idiom for GCP SM
Common idiom for GCP SM is like:
{
    "name": {
        "client_id": "<client_id>",
        "client_secret": "<client_secret>",
    }
}

Using DataFrom, the current GCP SM will fail to unmarshal cause it's expecting
{
    "name": "{\"client_id\": \"<client_id>\", \"client_secret\": \"<client_secret>\"}",
}
which is much more annoying to work with.  By tweaking the implementation to
try to decode to a string, but if that fails, use the raw []byte of the value
we get the best of both worlds.