deps: remove unncessary replace statements
PR #1990 attempted to bump the version of some dependencies but missed
the versions being set in the replace statements. This caused some of
the deps to not actually get updated (as can be seen by the contents of
the go.sum file). It turns out most of these replace statements are for
libraries that aren't currently being imported, so I cleaned up the
whole block.

The resulting changes can be seen in the go.sum file

Signed-off-by: Steve Mitchell <steve@sgmitchell.net>