go.mod dependency cleanup (#2085)
* deps: remove awkward k8s.io/client-go version

PR #1525 accidentally assumed that k8s.io/client-go followed semvar and
update the lib to the latest 1.x release. Unfortunately, that project
doesn't follow semvar on major versions so this actually _downgraded_
the package to one ~15 months earlier. This was subsequently fixed with
replace statements but the go mod file is easier to reason about if we
correct this

Signed-off-by: Steve Mitchell <steve@sgmitchell.net>

* deps: remove unncessary replace statements

PR #1990 attempted to bump the version of some dependencies but missed
the versions being set in the replace statements. This caused some of
the deps to not actually get updated (as can be seen by the contents of
the go.sum file). It turns out most of these replace statements are for
libraries that aren't currently being imported, so I cleaned up the
whole block.

The resulting changes can be seen in the go.sum file

Signed-off-by: Steve Mitchell <steve@sgmitchell.net>

---------

Signed-off-by: Steve Mitchell <steve@sgmitchell.net>