fix: chart: update cert-manager cert. duration (#2986)
Update cert-manager certificate duration to 1 year in the Helm chart.

This commit resolves a timing issue in the external-secrets Helm chart,
where the default certificate duration was previously not explicitly
set. This lack of specification led to conflicts with the cert-manager's
lookahead interval. By setting the `webhook.certManager.cert.duration`
to "8760h" (one year), we ensure that cert-manager will renew the
certificate before the external-secrets webhook starts to report issues,
and restarts, due to the certificate nearing expiration (as per the
lookahead interval).

This solution has been discussed in
external-secrets/external-secrets#2519.

Signed-off-by: Thibault Gérondal <tgerondal@emasphere.com>