fix(vault): Fix crash when caching is enabled and a token expires (#3598)
In the vault client library, LookupSelfWithContext calls ParseSecret,
which has a few places where it returns `nil, nil` instead of returning
a proper error. The most common scenario is when the token expires and
the Vault server returns:

    {
      "errors": [
        "permission denied"
      ]
    }

This commit adds an additional check to ensure that a nil response won't
be dereferenced in checkToken().

Signed-off-by: Andrew Gunnerson <andrew.gunnerson@elastic.co>