Ignore reconciliation errors caused by namespace termination
When namespace finalizers are used, it's pretty easy to hit a race
condition where a namespace is deleted while the ClusterExternalSecrets
and ExternalSecrets reconciliation loops are running. This commit
excludes errors caused by namespace deletion from being logged or
counting towards the error metrics since they are not actionable and are
"expected".

Signed-off-by: Andrew Gunnerson <andrew.gunnerson@elastic.co>