fix: improve webhook provider PushSecret handling
Refactor Webhook provider so that the body can be specified as a
template. This allows a secret to be sent to a web provider without
requiring the web provider to accept the secret in whatever form the
secret itself is in; the secret could be provided in a well-formed,
provider-specific JSON blob.

Signed-off-by: Billie Cleek <billie.cleek@lambdal.com>