fix: improve webhook provider PushSecret handling (#4508)
* fix: improve webhook provider PushSecret handling

Refactor Webhook provider so that the body can be specified as a
template. This allows a secret to be sent to a web provider without
requiring the web provider to accept the secret in whatever form the
secret itself is in; the secret could be provided in a well-formed,
provider-specific JSON blob.

Signed-off-by: Billie Cleek <billie.cleek@lambdal.com>

* maintain backward compatibility

Keep backward compatibility by sending the secret in the body when the
webhook provider's body field is empty.

Signed-off-by: Billie Cleek <billie.cleek@lambdal.com>

* docs: clarify Webhook PushSecret capability

Clarify that the webhook provider makes the secret available on the
remoteRef object so that it can be used in templates.

Signed-off-by: Billie Cleek <billie.cleek@lambdal.com>

* docs: clarify Webhook empty body capability

Clarify how to send an empty body with the Webhook provider pushing a
secret.

Signed-off-by: Billie Cleek <billie.cleek@lambdal.com>

---------

Signed-off-by: Billie Cleek <billie.cleek@lambdal.com>
Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>