feat: Get AWS secretsManager secrets using tag key value pairs.
The previous implementation matched any key and any value. This was behaviour was a surprise to our team.
Most systems that allow you to find resources by tags or labels do some filtering by key value pair.

After calling the listSecrets API with tags, we filter down the list to ensure that they match the key value pairs defined in the externalSecret.

Signed-off-by: Devon Hakel-Kinko <10181884+devonhk@users.noreply.github.com>