Update full-cluster-secret-store example to reference secret namespaces (#4964)
The example as written doesn't work if you're trying to create an external secret in any namespace other than the one with your AWS access key secrets in it. We should specify the namespace of the secrets in the ClusterSecretStore example so that it actually works for people following along when creating ExternalSecrets in another namespace that reference the cluster-wide store.

Signed-off-by: Jaz <ericvolp12@gmail.com>
Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>