docs: update bitwarden cli Dockerfile example
i've been maintaining a pull-able image of the example `Dockerfile`
since Jan 2024, when i discovered that the previous one was no longer
available.

in that time, i've gotten some PRs to my own repo with people proposing
changes, including validating the sha256sum of the `bw.zip` artifact.

github recently added a feature that provides a sha256sum of all release
artifacts, so bitwarden no longer publishes their own  checksums as txt
files.

this PR updates the example `Dockerfile` for the bitwarden cli to match
the one in the repo i'm maintaining to include:
1) handling the new checksum logic
2) running as non-root user
3) some minor tweaks to make the final container smaller & runnable as
   non-root

refs:
    - [the `Dockerfile` in the example repo](https://github.com/charlesthomas/bitwarden-cli/blob/2025.6.1/Dockerfile)
    - [the repo where i'm running the example image](https://github.com/charlesthomas/homelab-external-secrets/commit/a9ebfe0bfb908fe53c43ee96694bed50eb6e675f)